[Git] Don't store future FLAC files on LFS
GitHub's LFS has an unexpected quota of 1 GB in the free tier that we
only barely didn't hit with P0269. Adding another 50 GB costs $5 a
month, and doesn't really make sense for a repo that's only there for
reference and accounting purposes and that I expect barely anyone to
ever check out.